fix: Moved labels used to display counts to the positions of appropriate rings.
The positions of the rings showing progress and their icons had been moved, but the labels showing the actual values of those counts had not.  These labels have now been moved so that, for example, the ring and icon for calories has the label showing the count of calories below.